+++ api/api_common.c
@@ -278,7 +278,7 @@ struct FXs *getFXsFromNum(int windownum,int blocknum,int tracknum,int fxnum){
   return getFXsFromNumA(windownum, &window, blocknum, &wblock, tracknum, &wtrack, fxnum);
 }
 
-struct Patch *getPatchFromNum(int instrument_id){
+struct Patch *getPatchFromNum(int64_t instrument_id){
   struct Patch *patch = PATCH_get_from_id(instrument_id);
   if(patch==NULL)
     GFX_Message(NULL, "instrument %d not found", instrument_id);
@@ -286,7 +286,7 @@ struct Patch *getPatchFromNum(int instrument_id){
   return patch;
 }
 
-struct Patch *getAudioPatchFromNum(int instrument_id){
+struct Patch *getAudioPatchFromNum(int64_t instrument_id){
   struct Patch *patch = PATCH_get_from_id(instrument_id);
   if(patch==NULL) {
     GFX_Message(NULL, "instrument %d not found", instrument_id);
